更新

+++ include/tinyToolkit/log/event.h
@@ -133,8 +133,8 @@ namespace tinyToolkit
 	public:
 		std::tm tm{ };
 
-		std::size_t threadID{ OS::ThreadID() };
-		std::size_t processID{ OS::ProcessID() };
+		std::size_t threadID{ static_cast<std::size_t>(OS::ThreadID()) };
+		std::size_t processID{ static_cast<std::size_t>(OS::ProcessID()) };
 
 		std::string name{ };
 		std::string message{ };
